main_cb_important/requirements.txt
Flask==2.3.3
Flasgger==0.9.7
google-generativeai==0.1.2
